requirements.txt
torch
torchvision
transformers==4.46.3
timm
einops
pillow
opencv-python